requirements.txt
torch
torchvision
timm
numpy
pillow
matplotlib
tqdm
scikit-learn
jupyter
ultralytics
torchmetrics
detectron2
